requirements.txt
google-generativeai
pydantic
rich
psutil
tenacity
requests
python-dotenv
playwright
pocketbase
pytest
pytest-asyncio